refactor: remove processInstanceDetailsDiagramStore dependency from instanceHistoryModificationStore
